move toJsonString() & toSpecCasePath() into utils.spec
