+++ test/clear.spec.ts
@@ -1,6 +1,6 @@
 import { clear } from "../src/index";
 import { load } from "../src/load";
-import { toJsonString, toSpecCasePath } from "./load.spec";
+import { toJsonString, toSpecCasePath } from "./utils.spec";
 import { DEFAULT_SPEC_CASE_FOLDER } from "./spec_constants";
 import { StoreTestResult } from "./store.spec";
 import { getElementPathInfo } from "../src/utils";
