+++ test/delete.spec.ts
@@ -1,5 +1,5 @@
 import { deleteElement, load, store } from "../src/index";
-import { toJsonString, toSpecCasePath } from "./load.spec";
+import { toJsonString, toSpecCasePath } from "./utils.spec";
 import { DEFAULT_SPEC_CASE_FOLDER } from "./spec_constants";
 import { StoreTestResult } from "./store.spec";
 import { getElementPathInfo } from "../src/utils";
